+++ src/dgisim/status/status.py
@@ -133,6 +133,7 @@ __all__ = [
     "WindAndFreedomStatus",
 
     # character status
+    "AdeptusTemptationStatus",
     "FrozenStatus",
     "JueyunGuobaStatus",
     "KingsSquireEffectStatus",
@@ -2822,6 +2823,41 @@ class WindAndFreedomStatus(CombatStatus):
 
 ############################## Character Status ##############################
 
+@dataclass(frozen=True, kw_only=True)
+class AdeptusTemptationStatus(CharacterStatus):
+    DMG_BOOST: ClassVar[int] = 3
+    REACTABLE_SIGNALS: ClassVar[frozenset[TriggeringSignal]] = frozenset((
+        TriggeringSignal.ROUND_END,
+    ))
+
+    @override
+    def _preprocess(
+            self,
+            game_state: GameState,
+            status_source: StaticTarget,
+            item: PreprocessableEvent,
+            signal: Preprocessables,
+    ) -> tuple[PreprocessableEvent, None | Self]:
+        if signal is Preprocessables.DMG_AMOUNT_PLUS:
+            assert isinstance(item, DmgPEvent)
+            dmg = item.dmg
+            if (
+                    dmg.source == status_source
+                    and dmg.damage_type.direct_elemental_burst()
+                    and dmg.damage_type.can_boost()
+            ):
+                return item.delta_damage(self.DMG_BOOST), None
+        return item, self
+
+    @override
+    def _react_to_signal(
+            self, game_state: GameState, source: StaticTarget, signal: TriggeringSignal
+    ) -> tuple[list[eft.Effect], None | Self]:
+        if signal is TriggeringSignal.ROUND_END:
+            return [], None
+        return [], self  # pragma: no cover
+
+
 @dataclass(frozen=True)
 class FrozenStatus(CharacterStatus):
     damage_boost: ClassVar[int] = 2
@@ -2852,7 +2888,7 @@ class FrozenStatus(CharacterStatus):
     @override
     def _react_to_signal(
             self, game_state: GameState, source: StaticTarget, signal: TriggeringSignal
-    ) -> tuple[list[eft.Effect], None | FrozenStatus]:
+    ) -> tuple[list[eft.Effect], None | Self]:
         if signal is TriggeringSignal.ROUND_END:
             return [], None
         return [], self  # pragma: no cover